# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bingo OroDig API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/balance:
    patch:
      operationId: updateUserBalance
      tags: [users]
      summary: Update user balance (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/block:
    patch:
      operationId: toggleBlockUser
      tags: [users]
      summary: Block/unblock user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List bingo rooms
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"

    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a bingo room (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found

    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List games
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"

    post:
      operationId: createGame
      tags: [games]
      summary: Create a game (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
  /games/{id}/control:
    post:
      operationId: controlGame
      tags: [games]
      summary: Control game state (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameControl"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/draw:
    post:
      operationId: drawNumber
      tags: [games]
      summary: Draw next ball (admin/auto)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawnNumber"
  /games/{id}/numbers:
    get:
      operationId: getDrawnNumbers
      tags: [games]
      summary: Get all drawn numbers for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrawnNumber"
  /games/{id}/winners:
    get:
      operationId: getGameWinners
      tags: [games]
      summary: Get winners for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Winner"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role, balance, isBlocked, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, moderator, player]
        balance:
          type: number
        isBlocked:
          type: boolean
        totalWins:
          type: integer
        totalGamesPlayed:
          type: integer
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [admin, moderator, player]
    BalanceUpdate:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        reason:
          type: string
    BlockToggle:
      type: object
      required: [isBlocked]
      properties:
        isBlocked:
          type: boolean
    Room:
      type: object
      required: [id, name, type, status, cardPrice, maxPlayers, ballInterval, prize, patternType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [classic, fast, vip, automatic]
        status:
          type: string
          enum: [active, inactive, playing]
        cardPrice:
          type: number
        maxPlayers:
          type: integer
        ballInterval:
          type: integer
        prize:
          type: number
        patternType:
          type: string
          enum: [line, fullCard, corners, x, diagonal]
        playerCount:
          type: integer
        currentGameId:
          type: ["integer", "null"]
        createdAt:
          type: string
    RoomInput:
      type: object
      required: [name, type, cardPrice, maxPlayers, ballInterval, prize, patternType]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [classic, fast, vip, automatic]
        cardPrice:
          type: number
        maxPlayers:
          type: integer
        ballInterval:
          type: integer
        prize:
          type: number
        patternType:
          type: string
          enum: [line, fullCard, corners, x, diagonal]
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [classic, fast, vip, automatic]
        status:
          type: string
          enum: [active, inactive, playing]
        cardPrice:
          type: number
        maxPlayers:
          type: integer
        ballInterval:
          type: integer
        prize:
          type: number
        patternType:
          type: string
          enum: [line, fullCard, corners, x, diagonal]
    Game:
      type: object
      required: [id, roomId, status, patternType, ballInterval, prize, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        status:
          type: string
          enum: [waiting, playing, paused, finished]
        patternType:
          type: string
          enum: [line, fullCard, corners, x, diagonal]
        ballInterval:
          type: integer
        prize:
          type: number
        winnerId:
          type: ["integer", "null"]
        winnerCardId:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    GameInput:
      type: object
      required: [roomId]
      properties:
        roomId:
          type: integer
        patternType:
          type: string
          enum: [line, fullCard, corners, x, diagonal]
        prize:
          type: number
        ballInterval:
          type: integer
    GameControl:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [start, pause, resume, finish, restart]
    DrawnNumber:
      type: object
      required: [id, gameId, number, letter, drawnAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        number:
          type: integer
        letter:
          type: string
        drawnAt:
          type: string
    Winner:
      type: object
      required: [id, gameId, userId, cardId, pattern, prize, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        userId:
          type: integer
        cardId:
          type: integer
        username:
          type: string
        pattern:
          type: string
        prize:
          type: number
        createdAt:
          type: string
...