# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bingo OroDig API
 * OpenAPI spec version: 0.1.0
 */
import type { GamePatternType } from './gamePatternType';
import type { GameStatus } from './gameStatus';

export interface Game {
  id: number;
  roomId: number;
  status: GameStatus;
  patternType: GamePatternType;
  ballInterval: number;
  prize: number;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerCardId?: number | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  createdAt: string;
}