# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bingo OroDig API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomPatternType } from './roomPatternType';
import type { RoomStatus } from './roomStatus';
import type { RoomType } from './roomType';

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  type: RoomType;
  status: RoomStatus;
  cardPrice: number;
  maxPlayers: number;
  ballInterval: number;
  prize: number;
  patternType: RoomPatternType;
  playerCount?: number;
  /** @nullable */
  currentGameId?: number | null;
  createdAt: string;
}